+++ my-app/src/components/hashtag/HashtagResultView/style.js
@@ -4,6 +4,7 @@ const Cards = styled.ul`
   display: flex;
   flex-direction: column;
   gap: 1.6rem;
+  margin-top: 6rem;
 `;
 
 export default Cards;
