+++ my-app/src/pages/Hashtag/HashtagResult/style.js
@@ -1,12 +1,14 @@
 import styled from 'styled-components';
 
-export const Container = styled.div`
-  padding: 4.8rem 0 6rem;
+// const Cards = styled.ul`
+//   display: flex;
+//   flex-direction: column;
+//   gap: 1.6rem;
+//   margin-top: 2.6rem;
+// `;
+
+const Container = styled.div`
+  padding: 4.8rem + 0 6rem;
 `;
 
-export const Cards = styled.ul`
-  display: flex;
-  flex-direction: column;
-  gap: 1.6rem;
-  margin-top: 2.6rem;
-`;
+export default Container;
